backend/requirements.txt
Django
djangorestframework
django-cors-headers
python-dotenv
Pillow
django-filter
django-extensions
openai
pandas
numpy
psycopg2-binary
djangorestframework-simplejwt
supabase
dj-database-url
openpyxl
gunicorn
whitenoise
pyotp
requests
pgvector
